pyproject.toml
[project]
name = "agentic-ai-learnloop-demo"
version = "0.1.0"
description = "LangChain Demo"
authors = [
    {name = "Prince Clumson-Eklu",email = "ameto888@protonmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14,<4.0.0"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-openai (>=1.1.12,<2.0.0)",
    "openai (>=2.29.0,<3.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=25.2.0,<26.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "bandit (>=1.9.4,<2.0.0)",
    "safety (>=3.7.0,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "drf-yasg (>=1.21.15,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
